requirements.txt
# DeepTrace — pinned dependencies
# Phase 1: all listed. Phase 2: all active.

# LangGraph / LangChain / LangSmith
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langsmith>=0.1.0

# LLM SDKs (Phase 1: imported only inside USE_MOCK=false branches)
anthropic==0.40.0
openai==1.50.0
google-generativeai==0.8.0

# Structured output
instructor==1.4.3
pydantic==2.9.0

# Neo4j
neo4j==5.26.0
networkx==3.4.0
pyvis==0.3.2

# Search
tavily-python==0.5.0

# Frontend
streamlit==1.41.0
plotly==5.24.0

# HTTP / async
aiohttp==3.11.0
tenacity>=8.4.1,<9.0.0

# PDF report
fpdf2==2.8.1

# Utilities
python-dotenv==1.0.1
rich==13.9.0
click==8.1.7

# Dev / test
pytest==8.3.0
pytest-asyncio==0.24.0
